compose cs  commit @pis0sion

+++ app/api/repositories/OrderRepositories.php
@@ -566,7 +566,7 @@ class OrderRepositories extends AbstractRepositories
     /**
      * getAgentIDByVarchar
      * @param string $agentsID
-     * @return false|int|string
+     * @return int
      */
     protected function getAgentIDByVarchar(string $agentsID)
     {
@@ -575,7 +575,7 @@ class OrderRepositories extends AbstractRepositories
         if ($agentsID == ",")
             return 0;
         $agentsIDArr = explode(",", $agentsID);
-        return end($agentsIDArr);
+        return (int)end($agentsIDArr);
     }
 
 }
